fixing logic of user table

+++ app/Controllers/UserManagementController.php
@@ -27,24 +27,43 @@ class UserManagementController implements IController {
         if(isset($_POST['action']) and $_POST['action'] == "delete"
             and isset($_POST['id_user'])
         ){
-            // provedu smazani uzivatele
-            $ok = $this->db->deleteUser(intval($_POST['id_user']));
-            if($ok){
-                $tplData['success'] = "OK";
-                $tplData['delete'] = "Uživatel s ID:$_POST[id_user] byl smazán z databáze.";
-            } else {
-                $tplData['success'] = "ERROR";
-                $tplData['delete'] = "Uživatele s ID:$_POST[id_user] se nepodařilo smazat z databáze.";
+            if($tplData['user'][0] == $_POST['id_user']) {
+                $tplData['success'] = "WARNING";
+                $tplData['delete'] = "Uživatel nesmí smazat sám sebe";
+            }else if($tplData['user'][1] == 2 && ($this->db->getUserRole($_POST['id_user']) == "Admin" || $this->db->getUserRole($_POST['id_user']) == "Super Admin")){ //test práva
+                $tplData['success'] = "WARNING";
+                $tplData['delete'] = "Adminstrátorské účty může mazat jenom Super Admin.";
+            }else if($this->db->haveUserReview($_POST['id_user'])){
+                $tplData['success'] = "WARNING";
+                $tplData['delete'] = "Uživatel s ID:$_POST[id_user] nelze smazat, jelikož zveřejnil recenze.<br>Pokud chcete uživatele smazat, budete muset nejdříve smazat jeho recenze.";
+            }else{
+                // provedu smazani uzivatele
+                $ok = $this->db->deleteUser(intval($_POST['id_user']));
+                if($ok){
+                    $tplData['success'] = "OK";
+                    $tplData['delete'] = "Uživatel s ID:$_POST[id_user] byl smazán z databáze.";
+                } else {
+                    $tplData['success'] = "ERROR";
+                    $tplData['delete'] = "Uživatele s ID:$_POST[id_user] se nepodařilo smazat z databáze.";
+                }
             }
         }else if(isset($_POST['pravo'])){
-            $user = $this->db->getUser($_POST['id_user']);
-            $update = $this->db->updateUser($_POST["id_user"],$user[0]["username"], $user[0]["heslo"], $user[0]["email"], $_POST["pravo"]);
-            if($update){
-                $tplData['success'] = "OK";
-                $tplData['delete'] = "Uživateli s ID:$_POST[id_user] se podařilo změnit právo.";
-            }else{
-                $tplData['success'] = "ERROR";
-                $tplData['delete'] = "Uživateli s ID:$_POST[id_user] se nepodařilo změnit právo.";
+            if($tplData['user'][0] == $_POST['id_user']){
+                $tplData['success'] = "WARNING";
+                $tplData['delete'] = "Uživatel nesmí měnit právo sám sobě.";
+            }else if($tplData['user'][1] == 2 && ($this->db->getUserRole($_POST['id_user']) == "Admin" || $this->db->getUserRole($_POST['id_user']) == "Super Admin" || $_POST["pravo"] == "1" || $_POST["pravo"] == "2")) {
+                $tplData['success'] = "WARNING";
+                $tplData['delete'] = "Admin role může dávat, nebo rušit jenom Super Admin.";
+            } else{
+                $user = $this->db->getUser($_POST['id_user']);
+                $update = $this->db->updateUser($_POST["id_user"], $user[0]["username"], $user[0]["heslo"], $user[0]["email"], $_POST["pravo"]);
+                if ($update) {
+                    $tplData['success'] = "OK";
+                    $tplData['delete'] = "Uživateli s ID:$_POST[id_user] se podařilo změnit právo.";
+                } else {
+                    $tplData['success'] = "ERROR";
+                    $tplData['delete'] = "Uživateli s ID:$_POST[id_user] se nepodařilo změnit právo.";
+                }
             }
         }
